requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.6
django-cors-headers==4.7.0
django-model-utils==5.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
inflection==0.5.1
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
